# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Riftbound: Echoes of the Aether API",
    "description": "Server-Authoritative Anti-Cheat & Live-Ops API for Riftbound: Echoes of the Aether",
    "version": "1.0.0-PROD"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/player/profile": {
      "get": {
        "tags": [
          "Player & Inventory"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_v1_player_profile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/player/inventory": {
      "get": {
        "tags": [
          "Player & Inventory"
        ],
        "summary": "Get Inventory",
        "operationId": "get_inventory_api_v1_player_inventory_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryItemResponse"
                  },
                  "title": "Response Get Inventory Api V1 Player Inventory Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/guild/my-guild": {
      "get": {
        "tags": [
          "Guild Citadel"
        ],
        "summary": "Get My Guild",
        "operationId": "get_my_guild_api_v1_guild_my_guild_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guild/create": {
      "post": {
        "tags": [
          "Guild Citadel"
        ],
        "summary": "Create Guild",
        "operationId": "create_guild_api_v1_guild_create_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuildCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guild/contribute": {
      "post": {
        "tags": [
          "Guild Citadel"
        ],
        "summary": "Contribute Aether",
        "operationId": "contribute_aether_api_v1_guild_contribute_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuildContributeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GuildContributeRequest": {
        "properties": {
          "aether_amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Aether Amount"
          }
        },
        "type": "object",
        "required": [
          "aether_amount"
        ],
        "title": "GuildContributeRequest"
      },
      "GuildCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 32,
            "minLength": 3,
            "title": "Name"
          },
          "tag": {
            "type": "string",
            "maxLength": 6,
            "minLength": 2,
            "title": "Tag"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "tag"
        ],
        "title": "GuildCreateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InventoryItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "item_type": {
            "type": "string",
            "title": "Item Type"
          },
          "tier": {
            "type": "integer",
            "title": "Tier"
          },
          "quality": {
            "type": "string",
            "title": "Quality"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "is_equipped": {
            "type": "boolean",
            "title": "Is Equipped"
          }
        },
        "type": "object",
        "required": [
          "id",
          "item_id",
          "item_type",
          "tier",
          "quality",
          "attributes",
          "quantity",
          "is_equipped"
        ],
        "title": "InventoryItemResponse"
      },
...
      "ProfileResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "account_level": {
            "type": "integer",
            "title": "Account Level"
          },
          "mastery_xp": {
            "type": "integer",
            "title": "Mastery Xp"
          },
          "aether_shards": {
            "type": "integer",
            "title": "Aether Shards"
          },
          "void_fragments": {
            "type": "integer",
            "title": "Void Fragments"
          },
          "astral_cores": {
            "type": "integer",
            "title": "Astral Cores"
          },
          "world_contribution_points": {
            "type": "integer",
            "title": "World Contribution Points"
          },
          "active_hero_id": {
            "type": "string",
            "title": "Active Hero Id"
          },
          "current_season_score": {
            "type": "integer",
            "title": "Current Season Score"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "avatar_frame_id": {
            "type": "string",
            "title": "Avatar Frame Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "account_level",
          "mastery_xp",
          "aether_shards",
          "void_fragments",
          "astral_cores",
          "world_contribution_points",
          "active_hero_id",
          "current_season_score",
          "title",
          "avatar_frame_id"
        ],
        "title": "ProfileResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "username"
        ],
        "title": "TokenResponse"
      },
      "UserLoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLoginRequest"
      },
      "UserRegisterRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "minLength": 3,
            "title": "Username"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserRegisterRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}